requirements.txt
# Core OCR and Image Processing
opencv-python>=4.8.0
easyocr>=1.7.0
pytesseract>=0.3.10
Pillow>=10.0.0

# PDF Processing
PyMuPDF>=1.23.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP Requests for LLM APIs
requests>=2.31.0

# Optional: GPU acceleration for EasyOCR (uncomment if you have CUDA)
# torch>=2.0.0
# torchvision>=0.15.0

# System dependencies (install separately):
# - Tesseract OCR: 
#   - macOS: brew install tesseract
#   - Ubuntu: sudo apt install tesseract-ocr
#   - Windows: Download from https://github.com/UB-Mannheim/tesseract/wiki